requirements.txt
django
django_extensions
djangorestframework
django-filter
markdown